requirements.txt
opencv-python
matplotlib
tensorflow
tensorflow-gpu
Pillow
numpy
setuptools
wheel

